simulator/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
numpy==1.26.4
pandas==2.2.0
python-dotenv==1.0.1
